Bump atree version to v0.12.0 and cbor version to latest
This commit bumps atree version to v0.12.0 and also bumps
fxamacker/cbor to stream-mode branch based on v2.9.0.

This commit also updates the following API to match the changes
introduced by atree v0.12:
- Old API: "Value.Storable(SlabStorage, Address, uint64) (Storable, error)"
  New API: "Value.Storable(SlabStorage, Address, uint32) (Storable, error)"
- Old API: "WrapperValue.UnwrapAtreeValue() (Value, uint64)"
  New API: "WrapperValue.UnwrapAtreeValue() (Value, uint32)"

The API changes reduce frequent type casting in both atree and cadence.
The changes in Atree v0.12 do not affect stored data.

+++ interpreter/uint64atreevalue.go
@@ -35,7 +35,7 @@ var _ atree.Storable = Uint64AtreeValue(0)
 func (v Uint64AtreeValue) Storable(
 	_ atree.SlabStorage,
 	_ atree.Address,
-	_ uint64,
+	_ uint32,
 ) (
 	atree.Storable,
 	error,
