+++ interpreter/value_character.go
@@ -206,7 +206,7 @@ func (v CharacterValue) ConformsToStaticType(
 	return true
 }
 
-func (v CharacterValue) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v CharacterValue) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
