+++ interpreter/value_int16.go
@@ -603,7 +603,7 @@ func (v Int16Value) ConformsToStaticType(
 	return true
 }
 
-func (v Int16Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v Int16Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
