+++ interpreter/value_int256.go
@@ -697,7 +697,7 @@ func (v Int256Value) ConformsToStaticType(
 	return true
 }
 
-func (v Int256Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v Int256Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
