+++ interpreter/value_int64.go
@@ -599,7 +599,7 @@ func (v Int64Value) ConformsToStaticType(
 	return true
 }
 
-func (v Int64Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v Int64Value) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
