+++ interpreter/value_placeholder.go
@@ -69,7 +69,7 @@ func (PlaceholderValue) ConformsToStaticType(
 	return true
 }
 
-func (v PlaceholderValue) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v PlaceholderValue) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return NonStorable{Value: v}, nil
 }
