+++ interpreter/value_storagecapabilitycontroller.go
@@ -176,7 +176,7 @@ func (*StorageCapabilityControllerValue) IsStorable() bool {
 func (v *StorageCapabilityControllerValue) Storable(
 	storage atree.SlabStorage,
 	address atree.Address,
-	maxInlineSize uint64,
+	maxInlineSize uint32,
 ) (
 	atree.Storable,
 	error,
