+++ interpreter/value_type.go
@@ -283,7 +283,7 @@ func (v TypeValue) ConformsToStaticType(
 func (v TypeValue) Storable(
 	storage atree.SlabStorage,
 	address atree.Address,
-	maxInlineSize uint64,
+	maxInlineSize uint32,
 ) (atree.Storable, error) {
 	return values.MaybeLargeImmutableStorable(
 		v,
